refactor: refactor utils submodule
